LayoutTests/media/modern-media-controls/media-controller/ios/media-controller-scale-factor-audio-expected.txt

Testing that we apply a CSS zoom to ensure controls are always 50 device points tall on iOS for audio and also center the controls vertically.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('.media-controls').style.width is "800px"
PASS shadowRoot.querySelector('.media-controls').style.height is "62px"
PASS shadowRoot.querySelector('.media-controls').style.transform is "scale(0.5)"
PASS mediaBounds.top is controlsBounds.top
PASS mediaBounds.right is controlsBounds.right
PASS mediaBounds.bottom is controlsBounds.bottom
PASS mediaBounds.left is controlsBounds.left

PASS successfullyParsed is true

TEST COMPLETE
